topページの記事表示機能実装、バックエンドのリビルド用に構文エラー修正

+++ backend/src/main/java/com/bridge/backend/service/UserService.java
@@ -250,7 +250,7 @@ public class UserService {
                 .orElseThrow(() -> new RuntimeException("User not found"));
         user.setIcon(photoId);
         userRepository.save(user);
-        return getUserById(userId.longValue());
+        return getUserById(userId);
     }
 
     // 希望業界の更新
